chore: update test
